chore(src): add source dir and move lingusit and microphone inside
